feat: Добавил немного документации в Swagger а так же тесты

+++ src/main/java/com/sabadon/dailycalories/dto/meal/MealResponse.java
@@ -12,7 +12,7 @@ public record MealResponse(
 ) {
     public record MealDishResponse(
             String name,
-            BigDecimal portionSize,
+            Long portionSize,
             BigDecimal calories
     ) {
     }
